Dialog and Particles
There is now a DialogPrompt prefab that has been placed over many of the objects in the scene. Whenever they are walked over, dialog is shown to the player telling them their objective. One of six sentences are displayed based on how many times the player has checked the object out. The terrarium has been added to the final scene and particles shoot out when it is reached. Right now the dialogue prefabs do not carry over from scene to scene while playing the game.